Move store to bin/glusterd2/store
