+++ bin/glusterd2/store/embed.go
