+++ bin/glusterd2/store/store.go
