+++ bin/glusterd2/transaction/lock.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/gluster/glusterd2/bin/glusterd2/gdctx"
-	"github.com/gluster/glusterd2/store"
+	"github.com/gluster/glusterd2/bin/glusterd2/store"
 
 	"github.com/coreos/etcd/clientv3/concurrency"
 	"github.com/pborman/uuid"
