+++ bin/glusterd2/transaction/transaction.go
@@ -6,7 +6,7 @@ import (
 	"expvar"
 	"fmt"
 
-	"github.com/gluster/glusterd2/store"
+	"github.com/gluster/glusterd2/bin/glusterd2/store"
 
 	"github.com/coreos/etcd/clientv3"
 	"github.com/pborman/uuid"
